New Enable Timers button
Some speakers want don't want their times visible on the prompter, but they'd like to know how long it took them to speak in front of the prompter, so they enable and disable timers as desired. For the total time to be computed correctly, clock operations, as they exist, need to be running at all times. Running clock operations at all times can result in degraded performance when running on low end hardware and while using screen projections. The button added on this commit allows the user to enable and disable timers from running in the background without the need for a timer to be visible at all times.