dsit: Compilable WASM build
This commit makes it possible to compile QPrompt for WASM.  It removes
the KCoreAddons dependency, leaving a gap, and depends on a single
tweak to Kirigami.  The following compiler flags must also be set:
-DWASM:BOOL=ON
-DKF_IGNORE_PLATFORM_CHECK:BOOL=ON
-DKCOREADDONS_USE_QML:BOOL=OFF
-DBUILD_SHARED_LIBS:BOOL=OFF
-DBUILD_TESTING:BOOL=OFF
-DUSE_DBUS:BOOL=OFF

Doesn't work with single-threaded WebAssembly, only multi-threaded.
Resulting build loads in Firefox but fails to run due to static build
being unable to locate KirigamiPlugin.  This is the same bug that
afflicts static MacOS builds, and probably all static builds but
Android.

The changes made here not only bring the WASM build closer to fruition,
but also Android and iOS builds as well.